+++ fe/src/Challenge/ChallengeView.js
@@ -1,30 +1,33 @@
-import { AmazonsLogic } from "../Games/Amazons/AmazonsLogic.js";
 import { observer } from "mobx-react";
-import RootStore from "../Store/RootStore.js";
 import React from "react";
+
+import { useRootStore } from "../Store/RootStore.js";
+
+import { AmazonsLogic } from "../Games/Amazons/AmazonsLogic.js";
+
 import { PlayerTagView } from "./Components/PlayerTagView";
 import { MoveList } from "./Components/MoveList";
 import { StatusIndicator } from "./Components/StatusIndicator";
 
-const ChallengeView = observer(({ challenge }) => {
+const ChallengeView = observer(({ challenge, move_handler }) => {
+  const RootStore = useRootStore();
   let return_content;
+
   if (challenge == null) {
     return_content = "Loading the game!";
   } else {
     let game_is_in_progress = challenge.status === "IN_PROGRESS";
-    let is_users_turn = challenge.game_state.turn === challenge.client_turn;
+    let is_users_turn =
+      challenge.game_state.turn ===
+      challenge.turn_of_user(RootStore.client_uid);
     let player_can_move = game_is_in_progress && is_users_turn;
-
-    let handle_move_fn = (move) => {
-      challenge.send_move(move);
-    };
 
     return_content = (
       <div className="challenge-wrapper">
         <div className="challenge-board">
           <challenge.ViewComponent
             game_state={challenge.game_state}
-            handle_move={player_can_move ? handle_move_fn : undefined}
+            handle_move={player_can_move ? move_handler : undefined}
             last_move={
               challenge.moves.length > 0
                 ? challenge.moves[challenge.moves.length - 1]
@@ -49,7 +52,7 @@ const ChallengeView = observer(({ challenge }) => {
               displayName={
                 challenge.players.length > 1 ? challenge.players[1] : null
               }
-              isSelf={challenge.client_turn === 1}
+              isSelf={challenge.turn_of_user(RootStore.client_uid) === 1}
               isTurn={challenge.game_state.turn === 1}
               colorBadge={challenge.game_state.turn_to_color[1].badge}
             />
@@ -66,7 +69,7 @@ const ChallengeView = observer(({ challenge }) => {
 
             <PlayerTagView
               displayName={challenge.players[0]}
-              isSelf={challenge.client_turn === 0}
+              isSelf={challenge.turn_of_user(RootStore.client_uid) === 0}
               isTurn={challenge.game_state.turn === 0}
               colorBadge={challenge.game_state.turn_to_color[0].badge}
             />