+++ fe/src/Challenge/ChallengeView.test.js
@@ -1,8 +1,17 @@
-import { ChallengeView } from "./ChallengeView";
-import { render, screen } from "@testing-library/react";
 import { Challenge } from "./Challenge";
+import { ChallengeView } from "./ChallengeView";
+
+import { render, screen } from "@testing-library/react";
+
+import { RootStoreProvider, initRootStoreAndSocket } from "../Store/RootStore";
+
 import { NEW_AMAZONS_CHALLENGE_DATA } from "./Challenge.test";
 import { AMAZONS_CHALLENGE_DATA_AFTER_MOVE } from "./Challenge.test";
+
+const createFreshRootStore = () => {
+  const RootStore = initRootStoreAndSocket();
+  return RootStore;
+};
 
 describe("ChallengeView", () => {
   it("renders text when it does not have challenge data", () => {
@@ -13,19 +22,27 @@ describe("ChallengeView", () => {
   });
 
   it("renders challenge with challenge data", () => {
-    const challenge = new Challenge(NEW_AMAZONS_CHALLENGE_DATA);
+    const fresh_challenge = new Challenge(NEW_AMAZONS_CHALLENGE_DATA);
+    const RootStore = createFreshRootStore();
 
     const { asFragment } = render(
-      <ChallengeView challenge={challenge}></ChallengeView>
+      <RootStoreProvider store={RootStore}>
+        <ChallengeView challenge={fresh_challenge}></ChallengeView>
+      </RootStoreProvider>
     );
     expect(asFragment()).toMatchSnapshot();
   });
 
   it("renders challenge with at least one move and the last move indicator", () => {
-    const challenge = new Challenge(AMAZONS_CHALLENGE_DATA_AFTER_MOVE);
+    const challenge_with_one_move = new Challenge(
+      AMAZONS_CHALLENGE_DATA_AFTER_MOVE
+    );
+    const RootStore = createFreshRootStore();
 
     const { asFragment } = render(
-      <ChallengeView challenge={challenge}></ChallengeView>
+      <RootStoreProvider store={RootStore}>
+        <ChallengeView challenge={challenge_with_one_move}></ChallengeView>
+      </RootStoreProvider>
     );
     expect(asFragment()).toMatchSnapshot();
   });