+++ fe/src/Store/RootStore.js
@@ -1,6 +1,7 @@
 import { makeObservable, observable, action } from "mobx";
 import { Challenge } from "../Challenge/Challenge.js";
 import Socket from "../Network/Socket.js";
+import React from "react";
 
 class _RootStore {
   challenges = observable.map({});
@@ -14,6 +15,7 @@ class _RootStore {
       set_client_uid: action,
       update_challenge_information: action,
     });
+    this.db = "SINAN";
   }
 
   set_socket = (sckt) => {
@@ -36,11 +38,22 @@ class _RootStore {
   };
 }
 
-// we create all of our stores, socket IO class
-// and export them from this module to make it available to access
-// from every file without using React Contexts
-const RootStore = new _RootStore();
-const socket = new Socket();
-RootStore.set_socket(socket);
+//  Store helpers from https://codingislove.com/setup-mobx-react-context/
+const RootStoreContext = React.createContext();
 
-export default RootStore;
+export const initRootStoreAndSocket = () => {
+  const RootStore = new _RootStore();
+  const socket = new Socket(RootStore);
+  RootStore.set_socket(socket);
+  return RootStore;
+};
+
+export const RootStoreProvider = ({ children, store }) => {
+  return (
+    <RootStoreContext.Provider value={store}>
+      {children}
+    </RootStoreContext.Provider>
+  );
+};
+
+export const useRootStore = () => React.useContext(RootStoreContext);